fix #36; perform profile validation only to first certificate on chain; make public the profileValidation
