--- transfer-manager/src/main/java/eu/europa/ec/eudi/iso18013/transfer/internal/readerauth/profile/ProfileValidationImpl.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright (c) 2023-2024 European Commission
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package eu.europa.ec.eudi.iso18013.transfer.internal.readerauth.profile
-
-import java.security.cert.X509Certificate
-
-internal class ProfileValidationImpl(private val profileValidations: Collection<ProfileValidation>) :
-    ProfileValidation {
-    override fun validate(
-        readerAuthCertificate: X509Certificate,
-        trustCA: X509Certificate,
-    ): Boolean {
-        return profileValidations.all {
-            it.validate(readerAuthCertificate, trustCA)
-        }
-    }
-}
