+++ transfer-manager/src/main/java/eu/europa/ec/eudi/iso18013/transfer/readerauth/profile/ProfileValidation.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2024 European Commission
+ * Copyright (c) 2023-2025 European Commission
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.iso18013.transfer.internal.readerauth.profile
+package eu.europa.ec.eudi.iso18013.transfer.readerauth.profile
 
 import java.security.cert.X509Certificate
 
-internal fun interface ProfileValidation {
+fun interface ProfileValidation {
 
     companion object {
         @JvmStatic
@@ -35,5 +35,9 @@ internal fun interface ProfileValidation {
             ),
         )
     }
-    fun validate(readerAuthCertificate: X509Certificate, trustCA: X509Certificate): Boolean
+
+    fun validate(
+        chain: List<X509Certificate>,
+        trustCA: X509Certificate
+    ): Boolean
 }
