+++ transfer-manager/src/main/java/eu/europa/ec/eudi/iso18013/transfer/readerauth/profile/ProfileValidationImpl.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2024 European Commission
+ * Copyright (c) 2023-2025 European Commission
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,22 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.iso18013.transfer.internal.readerauth.profile
+package eu.europa.ec.eudi.iso18013.transfer.readerauth.profile
 
-import android.util.Log
-import eu.europa.ec.eudi.iso18013.transfer.internal.TAG
 import java.security.cert.X509Certificate
 
-private const val READER_AUTH_OID = "1.0.18013.5.1.6"
-
-internal class KeyExtended : ProfileValidation {
-
+class ProfileValidationImpl(private val profileValidations: Collection<ProfileValidation>) :
+    ProfileValidation {
     override fun validate(
-        readerAuthCertificate: X509Certificate,
+        chain: List<X509Certificate>,
         trustCA: X509Certificate,
     ): Boolean {
-        return readerAuthCertificate.extendedKeyUsage?.contains(READER_AUTH_OID).also {
-            Log.d(this.TAG, "KeyExtendedKeyUsage: $it")
-        } ?: false
+        return profileValidations.all {
+            it.validate(chain, trustCA)
+        }
     }
 }
