+++ transfer-manager/src/test/java/eu/europa/ec/eudi/iso18013/transfer/readerauth/profile/CommonNameTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2024 European Commission
+ * Copyright (c) 2023-2025 European Commission
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,26 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.iso18013.transfer.internal.readerauth.profile
+package eu.europa.ec.eudi.iso18013.transfer.readerauth.profile
 
 import android.util.Log
 import eu.europa.ec.eudi.iso18013.transfer.mockAndroidLog
 import eu.europa.ec.eudi.iso18013.transfer.readerauth.loadCert
 import eu.europa.ec.eudi.iso18013.transfer.readerauth.loadTrustCert
-import org.bouncycastle.asn1.x509.Extension.policyMappings
 import org.junit.After
 import org.junit.Before
 import org.mockito.MockedStatic
-import org.mockito.Mockito.mock
-import org.mockito.Mockito.`when`
+import org.mockito.Mockito
 import java.security.cert.X509Certificate
+import javax.security.auth.x500.X500Principal
 import kotlin.test.Test
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
-class CriticalExtensionsTest {
+class CommonNameTest {
 
-    private lateinit var readerAuthCertificate: X509Certificate
+    private lateinit var readerAuthCertificate: List<X509Certificate>
     private lateinit var trustCA: X509Certificate
 
     private lateinit var validation: ProfileValidation
@@ -41,9 +40,9 @@ class CriticalExtensionsTest {
 
     @Before
     fun setup() {
-        readerAuthCertificate = loadCert()
+        readerAuthCertificate = listOf(loadCert())
         trustCA = loadTrustCert()
-        validation = CriticalExtensions()
+        validation = CommonName()
 
         mockLog = mockAndroidLog()
     }
@@ -59,12 +58,14 @@ class CriticalExtensionsTest {
 
     @Test
     fun testVerify_Invalid() {
-        val mockCert = mock(X509Certificate::class.java)
-
-        `when`(mockCert.criticalExtensionOIDs).thenReturn(setOf<String>(policyMappings.id))
+        val mockCert = Mockito.mock(X509Certificate::class.java)
+        val principal = Mockito.mock(X500Principal::class.java).apply {
+            Mockito.`when`(name).thenReturn("")
+        }
+        Mockito.`when`(mockCert.subjectX500Principal).thenReturn(principal)
 
         // Call the method under test
-        val result = validation.validate(mockCert, trustCA)
+        val result = validation.validate(listOf(mockCert), trustCA)
 
         // Assert the result
         assertFalse(result)
