+++ transfer-manager/src/test/java/eu/europa/ec/eudi/iso18013/transfer/readerauth/profile/CriticalExtensionsTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2024 European Commission
+ * Copyright (c) 2023-2025 European Commission
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,24 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.iso18013.transfer.internal.readerauth.profile
+package eu.europa.ec.eudi.iso18013.transfer.readerauth.profile
 
 import android.util.Log
 import eu.europa.ec.eudi.iso18013.transfer.mockAndroidLog
 import eu.europa.ec.eudi.iso18013.transfer.readerauth.loadCert
-import eu.europa.ec.eudi.iso18013.transfer.readerauth.loadInvalidCert
 import eu.europa.ec.eudi.iso18013.transfer.readerauth.loadTrustCert
+import org.bouncycastle.asn1.x509.Extension.policyMappings
 import org.junit.After
 import org.junit.Before
 import org.mockito.MockedStatic
+import org.mockito.Mockito.mock
+import org.mockito.Mockito.`when`
 import java.security.cert.X509Certificate
 import kotlin.test.Test
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
-class PeriodTest {
+class CriticalExtensionsTest {
 
-    private lateinit var readerAuthCertificate: X509Certificate
+    private lateinit var readerAuthCertificate: List<X509Certificate>
     private lateinit var trustCA: X509Certificate
 
     private lateinit var validation: ProfileValidation
@@ -39,9 +41,9 @@ class PeriodTest {
 
     @Before
     fun setup() {
-        readerAuthCertificate = loadCert()
+        readerAuthCertificate = listOf(loadCert())
         trustCA = loadTrustCert()
-        validation = Period()
+        validation = CriticalExtensions()
 
         mockLog = mockAndroidLog()
     }
@@ -57,10 +59,12 @@ class PeriodTest {
 
     @Test
     fun testVerify_Invalid() {
-        val invalidCert = loadInvalidCert()
+        val mockCert = mock(X509Certificate::class.java)
+
+        `when`(mockCert.criticalExtensionOIDs).thenReturn(setOf<String>(policyMappings.id))
 
         // Call the method under test
-        val result = validation.validate(invalidCert, trustCA)
+        val result = validation.validate(listOf(mockCert), trustCA)
 
         // Assert the result
         assertFalse(result)
